+++ cypress/e2e/agreements/agreement-notes/create-agreement-notes.cy.js
@@ -12,48 +12,50 @@ let agreementId;
 let noteTypeId;
 const noteType = `NoteType ${randomFourDigitNumber()}`;
 
-describe('Agreement Notes', () => {
-  before('create test data', () => {
-    cy.getAdminToken();
-    AgreementsApi.createViaApi().then((agreement) => {
-      agreementId = agreement.id;
+describe('Agreements', () => {
+  describe('Agreement Notes', () => {
+    before('create test data', () => {
+      cy.getAdminToken();
+      AgreementsApi.createViaApi().then((agreement) => {
+        agreementId = agreement.id;
+      });
+      Notes.createNoteTypeViaApi(noteType).then((note) => {
+        noteTypeId = note.id;
+      });
+      cy.loginAsAdmin({ path: TopMenu.agreementsPath, waiter: Agreements.waitLoading });
     });
-    Notes.createNoteTypeViaApi(noteType).then((note) => {
-      noteTypeId = note.id;
+
+    after('delete test data', () => {
+      AgreementsApi.deleteViaApi(agreementId);
+      Notes.deleteNoteTypeViaApi(noteTypeId);
     });
-    cy.loginAsAdmin({ path: TopMenu.agreementsPath, waiter: Agreements.waitLoading });
+
+    it(
+      'C1308 Create a note for an Agreement record (erm)',
+      { tags: [TestTypes.extendedPath, DevTeams.erm] },
+      () => {
+        AgreementViewDetails.agreementListClick(AgreementsApi.defaultAgreement.name);
+        AgreementViewDetails.openNotesSection();
+        AgreementViewDetails.verifyNotesIsEmpty();
+
+        AgreementViewDetails.addNewNote();
+        NewNote.verifyNewNoteIsDisplayed();
+
+        NewNote.chooseSelectTypeByTitle(noteType);
+        NewNote.fill();
+        NewNote.save();
+        AgreementViewDetails.verifyAgreementDetailsIsDisplayedByTitle(
+          AgreementsApi.defaultAgreement.name,
+        );
+        AgreementViewDetails.verifyNotesCount('1');
+
+        AgreementViewDetails.openNotesSection();
+        AgreementViewDetails.verifySpecialNotesRow(
+          NewNote.defaultNote.title,
+          NewNote.defaultNote.details,
+          noteType,
+        );
+      },
+    );
   });
-
-  after('delete test data', () => {
-    AgreementsApi.deleteViaApi(agreementId);
-    Notes.deleteNoteTypeViaApi(noteTypeId);
-  });
-
-  it(
-    'C1308 Create a note for an Agreement record (erm)',
-    { tags: [TestTypes.extendedPath, DevTeams.erm] },
-    () => {
-      AgreementViewDetails.agreementListClick(AgreementsApi.defaultAgreement.name);
-      AgreementViewDetails.openNotesSection();
-      AgreementViewDetails.verifyNotesIsEmpty();
-
-      AgreementViewDetails.addNewNote();
-      NewNote.verifyNewNoteIsDisplayed();
-
-      NewNote.chooseSelectTypeByTitle(noteType);
-      NewNote.fill();
-      NewNote.save();
-      AgreementViewDetails.verifyAgreementDetailsIsDisplayedByTitle(
-        AgreementsApi.defaultAgreement.name,
-      );
-      AgreementViewDetails.verifyNotesCount('1');
-
-      AgreementViewDetails.openNotesSection();
-      AgreementViewDetails.verifySpecialNotesRow(
-        NewNote.defaultNote.title,
-        NewNote.defaultNote.details,
-        noteType,
-      );
-    },
-  );
 });
