--- cypress/e2e/eholdings/eholdings-packages.cy.js
@@ -1,179 +1,0 @@
-import testType from '../../support/dictionary/testTypes';
-import TopMenu from '../../support/fragments/topMenu';
-import eHoldingsPackages from '../../support/fragments/eholdings/eHoldingsPackages';
-import eHoldingSearch from '../../support/fragments/eholdings/eHoldingsSearch';
-import eHoldingsPackagesSearch from '../../support/fragments/eholdings/eHoldingsPackagesSearch';
-import eHoldingsTitlesSearch from '../../support/fragments/eholdings/eHoldingsTitlesSearch';
-import eHoldingsPackage from '../../support/fragments/eholdings/eHoldingsPackage';
-import permissions from '../../support/dictionary/permissions';
-import features from '../../support/dictionary/features';
-import users from '../../support/fragments/users/users';
-import devTeams from '../../support/dictionary/devTeams';
-import eHoldingsProvidersSearch from '../../support/fragments/eholdings/eHoldingsProvidersSearch';
-
-describe('eHoldings packages management', () => {
-  let userId;
-
-  afterEach(() => {
-    users.deleteViaApi(userId);
-  });
-
-  it(
-    'C688 Add all titles in a package to your holdings (spitfire)',
-    { tags: [testType.smoke, devTeams.spitfire, features.eHoldings] },
-    () => {
-      cy.createTempUser([
-        permissions.uieHoldingsRecordsEdit.gui,
-        permissions.uieHoldingsPackageTitleSelectUnselect.gui,
-        permissions.moduleeHoldingsEnabled.gui,
-      ]).then((userProperties) => {
-        userId = userProperties.userId;
-        eHoldingsPackages.getNotSelectedPackageIdViaApi().then((specialPackage) => {
-          cy.login(userProperties.username, userProperties.password, {
-            path: `${TopMenu.eholdingsPath}/packages/${specialPackage.id}`,
-            waiter: () => eHoldingsPackage.waitLoading(specialPackage.name),
-          });
-          eHoldingsPackage.addToHoldings();
-          eHoldingsPackage.verifyHoldingStatus();
-          eHoldingsPackage.filterTitles();
-          eHoldingsPackage.checkEmptyTitlesList();
-          // reset test data
-          eHoldingsPackage.removeFromHoldings();
-        });
-      });
-    },
-  );
-
-  it(
-    'C3463 Add two tags to package [Edinburgh Scholarship Online] (spitfire)',
-    { tags: [testType.smoke, devTeams.spitfire, features.eHoldings, features.tags] },
-    () => {
-      // TODO: "Tags: All permissions" doesn't have displayName. It's the reason why there is related permission name in response, see https://issues.folio.org/browse/UITAG-51
-      cy.createTempUser([
-        permissions.uieHoldingsRecordsEdit.gui,
-        permissions.uiTagsPermissionAll.gui,
-      ]).then((userProperties) => {
-        userId = userProperties.userId;
-        cy.login(userProperties.username, userProperties.password, {
-          path: TopMenu.eholdingsPath,
-          waiter: eHoldingsTitlesSearch.waitLoading,
-        });
-        eHoldingSearch.switchToPackages();
-        eHoldingsPackagesSearch.byName();
-        eHoldingsPackages.openPackage().then((selectedPackage) => {
-          const addedTag1 = eHoldingsPackage.addTag();
-          const addedTag2 = eHoldingsPackage.addTag();
-          eHoldingsPackage.close(selectedPackage);
-          eHoldingsPackagesSearch.byName(selectedPackage);
-          eHoldingsPackages.openPackage();
-          eHoldingsPackage.verifyExistingTags(addedTag1, addedTag2);
-        });
-      });
-    },
-  );
-
-  it(
-    'C3464 Update package proxy (spitfire)',
-    { tags: [testType.criticalPath, devTeams.spitfire, features.eHoldings] },
-    () => {
-      cy.createTempUser([permissions.uieHoldingsRecordsEdit.gui]).then((userProperties) => {
-        userId = userProperties.userId;
-        cy.login(userProperties.username, userProperties.password, {
-          path: TopMenu.eholdingsPath,
-          waiter: eHoldingsPackages.waitLoading,
-        });
-
-        eHoldingSearch.switchToPackages();
-        eHoldingsProvidersSearch.byProvider('Edinburgh Scholarship Online');
-        eHoldingsPackages.openPackage();
-        eHoldingsPackage.editProxyActions();
-        eHoldingsPackages.changePackageRecordProxy().then((newProxy) => {
-          eHoldingsPackage.saveAndClose();
-          // additional delay related with update of proxy information in ebsco services
-          cy.wait(10000);
-          eHoldingsPackages.checkPackageRecordProxy(newProxy);
-        });
-      });
-    },
-  );
-
-  it(
-    'C690 Remove a package from your holdings (spitfire)',
-    { tags: [testType.smoke, devTeams.spitfire, features.eHoldings] },
-    () => {
-      cy.createTempUser([
-        permissions.uieHoldingsRecordsEdit.gui,
-        permissions.uieHoldingsPackageTitleSelectUnselect.gui,
-        permissions.uieHoldingsTitlesPackagesCreateDelete.gui,
-      ]).then((userProperties) => {
-        userId = userProperties.userId;
-        eHoldingsPackages.getNotCustomSelectedPackageIdViaApi().then((specialPackage) => {
-          cy.login(userProperties.username, userProperties.password, {
-            path: `${TopMenu.eholdingsPath}/packages/${specialPackage.id}`,
-            waiter: () => eHoldingsPackage.waitLoading(specialPackage.name),
-          });
-          eHoldingsPackage.removeFromHoldings();
-          eHoldingsPackage.verifyHoldingStatus(eHoldingsPackage.filterStatuses.notSelected);
-          eHoldingsPackage.filterTitles(eHoldingsPackage.filterStatuses.notSelected);
-          eHoldingsPackage.checkEmptyTitlesList();
-          // reset test data
-          eHoldingsPackage.addToHoldings();
-        });
-      });
-    },
-  );
-
-  it(
-    'C695 Package Record: Search all titles included in a package (spitfire)',
-    { tags: [testType.criticalPath, devTeams.spitfire, features.eHoldings] },
-    () => {
-      cy.createTempUser([permissions.uieHoldingsRecordsEdit.gui]).then((userProperties) => {
-        userId = userProperties.userId;
-        cy.login(userProperties.username, userProperties.password, {
-          path: TopMenu.eholdingsPath,
-          waiter: eHoldingsPackages.waitLoading,
-        });
-
-        eHoldingSearch.switchToPackages();
-        eHoldingsProvidersSearch.byProvider('Wiley Online Library');
-        eHoldingsPackagesSearch.bySelectionStatus('Selected');
-        eHoldingsPackages.openPackage();
-        eHoldingsPackages.titlesSearch('Subject', 'engineering');
-        eHoldingsPackages.clickSearchTitles();
-        eHoldingsPackages.subjectsAssertion();
-      });
-    },
-  );
-
-  it(
-    'C756 Remove a tag from a package record (spitfire)',
-    { tags: [testType.extendedPath, devTeams.spitfire, features.eHoldings, features.tags] },
-    () => {
-      cy.createTempUser([
-        permissions.uieHoldingsRecordsEdit.gui,
-        permissions.uiTagsPermissionAll.gui,
-        permissions.uieHoldingsTitlesPackagesCreateDelete.gui,
-      ]).then((userProperties) => {
-        userId = userProperties.userId;
-        cy.login(userProperties.username, userProperties.password, {
-          path: TopMenu.eholdingsPath,
-          waiter: eHoldingsTitlesSearch.waitLoading,
-        });
-        eHoldingSearch.switchToPackages();
-        eHoldingsPackagesSearch.byName();
-        eHoldingsPackages.openPackage().then((selectedPackageName) => {
-          // existing test data clearing
-          eHoldingsPackage.removeExistingTags();
-          const addedTag = eHoldingsPackage.addTag();
-          eHoldingsPackage.close(selectedPackageName);
-          eHoldingsPackagesSearch.byTag(addedTag);
-          eHoldingsPackages.openPackage();
-          eHoldingsPackage.verifyExistingTags(addedTag);
-          eHoldingsPackage.removeExistingTags();
-          cy.reload();
-          eHoldingsPackage.verifyExistingTags();
-        });
-      });
-    },
-  );
-});
